quaternion-transformers/requirements.txt
tensorflow-gpu==2.4.1
tensorflow-datasets==4.2.0
tensorflow-text==2.4.3
tabulate==0.8.9
nltk==3.6.1